Fix release procedure of QueryDesc object from stack
The global state variable that denotes a stack of currently executing
queries requires a special care on supporting. The mechanism of adding
and removing of current stack frame in Executor hooks is kept from
auto_explain extension for nesting_level variable. The current stack
frame is added to stack before ExecutorRun and ExecutorFinish routines
and is released after them