[PGPRO-8358] Added the progress value to the pg_query_state message
The progress value is equal to the actual_rows ratio to plan_rows as a percentage.
Or 100% if actual_rows > plan_rows. The progress value is added for each node in the current_loop.
The progress value is added to the message taking into account the format (TEXT, XML,
JSON, YAML) and provided that the parameter costs = true.

Tags: pg_query_state